feat: add org page header and move hooks
- add header with org for all child pages
- move reused hooks into functions
