+++ src/app/[organizationId]/page.tsx
@@ -1,160 +1,23 @@
 'use client'
 
-import { ChevronRightIcon } from '@primer/octicons-react'
 import { ActionList, Avatar, Box, Flash, Link, Octicon } from '@primer/react'
-import { personalOctokit } from '../../bot/octokit'
 import { useParams } from 'next/navigation'
-import { FC, useEffect, useState } from 'react'
 import { trpc } from '../../utils/trpc'
 
 import { useSession } from 'next-auth/react'
+import { useOrgData } from 'utils/organization'
+import { useForksData } from 'utils/forks'
+import { RepoListItem } from 'app/components/RepoListItem'
 
-type ArrayElement<ArrayType extends readonly unknown[]> =
-  ArrayType extends readonly (infer ElementType)[] ? ElementType : never
-
-const getOrganizationInformation = async (
-  accessToken: string,
-  orgId: string,
-) => {
-  return (await personalOctokit(accessToken).rest.orgs.get({ org: orgId })).data
-}
-
-const getForksInOrg = async (accessToken: string, owner: string) => {
-  return (
-    await personalOctokit(accessToken).rest.repos.listForOrg({
-      per_page: 100,
-      type: 'forks',
-      org: owner,
-    })
-  ).data
-}
-
-const getParent = async (accessToken: string, owner: string, repo: string) => {
-  return (
-    await personalOctokit(accessToken).rest.repos.get({
-      owner,
-      repo,
-    })
-  ).data.parent
-}
-
-const RepoListItem: FC<{
-  fork: ArrayElement<Awaited<ReturnType<typeof getForksInOrg>>>
-  orgId: string
-  accessToken: string
-}> = ({ fork, orgId, accessToken }) => {
-  const [parent, setParent] = useState<Awaited<
-    ReturnType<typeof getParent>
-  > | null>(null)
-
-  useEffect(() => {
-    if (!fork) {
-      return
-    }
-
-    getParent(accessToken, fork.owner.login, fork.name).then((parent) => {
-      setParent(parent)
-    })
-  }, [fork, accessToken])
-
-  if (!fork) {
-    return null
-  }
-
-  return (
-    <ActionList.LinkItem
-      key={fork.id}
-      href={`/${orgId}/forks/${fork.id}`}
-      sx={{
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-      }}
-    >
-      <ActionList.LeadingVisual>
-        <Avatar src={parent?.owner.avatar_url ?? fork.owner.avatar_url} />
-      </ActionList.LeadingVisual>
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          ml: 2,
-        }}
-      >
-        <Box
-          sx={{
-            fontWeight: 600,
-            fontSize: 2,
-            mb: 1,
-          }}
-        >
-          {fork.owner.login}/{fork.name}
-        </Box>
-        <Box as="small">
-          <Box>
-            Forked from{' '}
-            {parent ? (
-              <Box
-                as="span"
-                sx={{
-                  fontWeight: 600,
-                }}
-              >
-                {parent.owner.login}/{parent.name}
-              </Box>
-            ) : (
-              '...'
-            )}
-          </Box>
-        </Box>
-      </Box>
-      <ActionList.TrailingVisual>
-        <Octicon icon={ChevronRightIcon} />
-      </ActionList.TrailingVisual>
-    </ActionList.LinkItem>
-  )
-}
-
-const SingleOrganization = () => {
+const Organization = () => {
   const session = useSession()
   const { accessToken } = (session.data?.user as any) ?? {}
   const { organizationId } = useParams()
-  const [orgData, setOrgData] = useState<Awaited<
-    ReturnType<typeof getOrganizationInformation>
-  > | null>(null)
-  const [forks, setForks] = useState<Awaited<
-    ReturnType<typeof getForksInOrg>
-  > | null>(null)
   const { data, isLoading } = trpc.checkInstallation.useQuery({
     orgId: organizationId as string,
   })
-
-  useEffect(() => {
-    if (!accessToken || !organizationId) {
-      return
-    }
-
-    getOrganizationInformation(accessToken, organizationId as string).then(
-      (orgData) => {
-        setOrgData(orgData)
-      },
-    )
-  }, [organizationId, accessToken])
-
-  useEffect(() => {
-    if (!orgData) {
-      return
-    }
-
-    getForksInOrg(accessToken, orgData.login).then((forks) => {
-      setForks(forks)
-    })
-  }, [orgData, accessToken])
-
-  // TODO: Handle org not found
-  if (!orgData) {
-    return <div>Loading organization data...</div>
-  }
+  const orgData = useOrgData()
+  const forks = useForksData()
 
   return (
     <Box>
@@ -170,23 +33,6 @@ const SingleOrganization = () => {
             to install the App to the organization.
           </Flash>
         )}
-      </Box>
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'row',
-          alignItems: 'center',
-        }}
-      >
-        <Avatar src={orgData.avatar_url} size={24} />
-        <Box
-          as="h3"
-          sx={{
-            ml: 2,
-          }}
-        >
-          {orgData.login}
-        </Box>
       </Box>
       <Box>
         {forks && (
@@ -207,4 +53,4 @@ const SingleOrganization = () => {
   )
 }
 
-export default SingleOrganization
+export default Organization
